Moved integration tests for creating an Itsystem and Tilstand to separate files according to pull request comment
